pyproject.toml
[project]
name = "Daily Papers Podcast"
version = "0.1.0"
description = "Pipeline for automatically generating podcasts about Arxiv papers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.18.0",
    "librosa>=0.11.0",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "soundfile>=0.13.1",
]

[dependency-groups]
dev = ["pyright>=1.1.401", "ruff>=0.11.12"]

[tool.ruff]
target-version = "py313"
line-length = 165
fix = true

[tool.ruff.lint]
select = ["ALL"]
ignore = ["COM812", "D"]
